+++ src/components/PaginationBase/PageItem.tsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import * as React from 'react';
 import { ReactNode } from 'react';
 import Anchor from '@restart/ui/Anchor';
-import { BsPrefixProps, BsPrefixRefForwardingComponent } from './helpers';
+import { BsPrefixProps, BsPrefixRefForwardingComponent } from '../helpers';
 
 export interface PageItemProps
   extends React.HTMLAttributes<HTMLElement>,
